+++ src/index.ts
@@ -16,3 +16,4 @@ export * from "./menu-dom.js";
 export * from "./tabs-dom.js";
 export * from "./dialog-dom.js";
 export * from "./popover-dom.js";
+export * from "./command-palette.js";